refactor!: Introduce versioning for proofs
Also:
- combine `proof_stream.rs` and `proof_item.rs` into `proof.rs`
- rename proof-related types
